requirements.txt
pdfplumber==0.11.0
PyMuPDF==1.24.3
pytesseract==0.3.13
pandas==2.2.2
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
streamlit==1.35.0  
scikit-learn==1.5.0     
thefuzz==0.22.1   
python-Levenshtein==0.25.1 
pytest==8.2.0
fpdf2==2.7.9